Move hypothesis tests
